// src/types/enum/ModerationPresetDuration.ts
import ms from "ms"

enum ModerationPresetDuration {
	THREE_DAYS = "THREE_DAYS",
	ONE_WEEK = "ONE_WEEK",
	TWO_WEEKS = "TWO_WEEKS",
	FOUR_WEEKS = "FOUR_WEEKS",
	TWELVE_WEEKS = "TWELVE_WEEKS",
	TWENTY_FOUR_WEEKS = "TWENTY_FOUR_WEEKS",
	ONE_YEAR = "ONE_YEAR",
	FOREVER = "FOREVER",
}

export default ModerationPresetDuration